+++ apps/app/navigation/screens/acceptWorkspaceInvitationScreen/AcceptWorkspaceInvitationScreen.tsx
@@ -18,7 +18,7 @@ import { LoginForm } from "../../../components/login/LoginForm";
 import { OnboardingScreenWrapper } from "../../../components/onboardingScreenWrapper/OnboardingScreenWrapper";
 import RegisterForm from "../../../components/register/RegisterForm";
 import { useWorkspaceInvitationQuery } from "../../../generated/graphql";
-import { RootStackScreenProps } from "../../../types/navigation";
+import { RootStackScreenProps } from "../../../types/navigationProps";
 import { acceptWorkspaceInvitation } from "../../../utils/workspace/acceptWorkspaceInvitation";
 
 const Wrapper = ({ children }) => (
