+++ apps/app/navigation/screens/sharePageScreen/SharePageScreen.tsx
@@ -1,5 +1,5 @@
 import { Heading } from "@serenity-tools/ui";
-import { RootStackScreenProps } from "../../../types/navigation";
+import { RootStackScreenProps } from "../../../types/navigationProps";
 
 export default function SharePageScreen(
   props: RootStackScreenProps<"SharePage">
